Log in with Password: Only match users with `password_bcrypt` (#466)
If a Project has only "Log in with Password" enabled and an end user
inputs an email/password pair matching a single user, but that user has
no `password_bcrypt`, then the user receives an error about no password
being registered.

This PR instead has our behavior be to fail to find a matching user for
the flow, and to fall back to the Log in with Email flow.