Console Redesign (#465)
This PR rewrites the entire Console UI from scratch. It reorganizes the
console pages into a more user-friendly hierarchy and completely
redesigned everything.

Since this is a full rewrite a number of the outstanding issues and
inconsistencies in the current Console are addressed.

The new console sitemap is broken down into `/settings` pages (all
Project configuration views) and `/organizations` pages (all
Organization and Organization sub-resources views).

![Screenshot 2025-06-17 at 3 09
26 PM](https://github.com/user-attachments/assets/3c031b2b-9bdb-4078-8a93-7e268ac09807)
![Screenshot 2025-06-17 at 3 19
14 PM](https://github.com/user-attachments/assets/0461995c-9167-4386-8e31-608d641c6b69)

---------

Co-authored-by: Ulysse Carion <ulysse.carion@ssoready.com>