Resolve issues with some list queries not selecting properly (#470)
The way that our page token logic works is that it encodes the next ID
in a query as the pageToken in the response. This is used in our list
queries (ordered by id) to `SELECT ... WHERE id >= @page_token`.

However, in some of our queries, we were selecting `WHERE id >
@page_token`. This resulted in single records being dropped when
paginating.

This PR updates these queries to select `WHERE id >= @page_token`
instead.