Add Sentry slog integration
Configures an slog handler for mirroring log records as Sentry breadcrumbs. If a Sentry event is triggered (either by panic or by calling sentry.CaptureException), all breadcrumbs up to that point will be attached to the event. The logic in sentryhttp and opaqueinternalerror ensure that every HTTP request has a unique context such that breadcrumbs are scoped to an HTTP request.