Add proactive access token refresh (#493)
Currently, if pages are left idle on the console or vault, the access
token becomes invalid and react query's automatic refetching will start
to 401.

The vault is currently polling on a 60-second cadence, which can result
in up to 50 seconds of expired access tokens. This PR updates both the
vault and console to poll on a 2 second interval to check if refresh is
required.