Add replica identity to enable cascading deletes of email quota records (#504)
Since there is no primary key on the `project_email_quota_daily_usage`
table, cascading deletes break as there is no replica identity.
Typically, with 1:1 tables (if we only had one record per project), we
could just make the project ID the primary key.

Since this is not the case, we need to use the full record as the
replica identity in order to postgres to be able to properly track
cascading deletes.

This PR alters the table to include this replica identity.