Fix passkeys (#506)
Currently, our Passkey Relying Party ID is operating on the assumptions
built up by the Google and Safari implementation of domain restriction
(which allow for subdomains of the current origin as a valid RP ID).
However, their implementations stray a bit from [the spec
itself](https://www.w3.org/TR/webauthn-2/#rp-id) which states that the
RP ID `must be equal to the origin's effective domain, or a registrable
domain suffix of the origin's effective domain`.

This PR updates our Passkey rpcs for issuing challenges and getting
options for Passkey registration and verification in both the frontend
and intermediate stores to use the Project's `cookie_domain` instead of
the `vault_domain`. This loosely ensures stronger compatibility with the
expectations of the spec (and solves some problems with Passkeys in
Firefox).

This has been confirmed working in Chrome, Firefox, and Safari both
locally and on dev1.