Add project setting to disable self-serve user creation (#538)
This PR adds an option to disable self-serve creation of users.

This setting is enforced in the Intermediate API's exchange flow
(preventing user creation) and in the Frontend API's user invite flow.

The Vault UI hides the User Invites functionality based on this setting.
The Vault login flow knows that when self-serve user creation is
disabled, only organizations that have `userExists` are valid to join.

By default, self-serve user creation is enabled. When disabled, the
Vault Behavior Settings card will also disable organization creation,
since in practice that flow won't work either.


![screenshot-2025-07-09-11-26-13](https://github.com/user-attachments/assets/ed99491f-8ad1-46b4-a343-67e1bb7f45f6)

This PR also makes a fix to the appearance of the user invite creation
card in the Vault. That's an opportunistic aside.