[OPEN-524] Remove .well-known service (#523)
This PR removes the `.well-known` service, since we're now using the
`cookie_domain` as the Relying Party ID for passkeys. With this in
place, we no longer need to support Apple's fallback to the
`.well-known` files to determine domain relationships, since the RP ID
is always expected to be in line with the spec and be either the current
domain or a registrable suffix of the current domain.