[OPEN-529] Remove redirect URI settings from console (#526)
This PR removes the Redirect URI settings from the vault details section
of the console. Since we're going to be removing these columns, we will
no longer need dedicated settings in the console for them.

The removal from the db, protos, and store layer will come in a separate
PR. This one is purely cosmetic.