[OPEN-530] Use cookie domain for default redirect URL in finish login (#525)
This PR updates the `finish-login` route in the vault to use the cookie
domain for the default redirect URI instead of the various redirect uri
settings (they will be getting removed soon).

In order to support this, it also updates the intermediate proto message
for `Settings` to include the `cookie_domain` and updates the store
function to include `qProject.CookieDomain` when parsing settings.